Solution to Logical OR Operation

+++ Logical OR Operation@@67882a731b4392d8fc1577dd/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
-float main(){
-    int a,b;
+int main(){
+    float a,b;
     scanf("%d %d",&a,&b);
     if(a||b){
         printf("True");
